+++ src/components/composite/DustFlow/hooks/useDustFormFields.ts
@@ -6,64 +6,31 @@ import {
   defineDisplayAmountField,
   defineComputedField,
 } from '@/components/composite/JumperWidget/utils';
-import type { ChainSingleSelectValue } from '@/components/composite/JumperWidget/components/Chain';
-import type { NumericSelectValue } from '@/components/composite/JumperWidget/components/NumericSelect';
-import type { BalancesMultiSelectValue } from '@/components/composite/JumperWidget/components/Balances';
-import type {
-  PortfolioBalance,
-  WalletToken,
-  ExtendedToken,
-} from '@/types/tokens';
+import type { ExtendedToken } from '@/types/tokens';
+import type { PortfolioBalance, WalletToken } from '@/types/tokens';
 import { useTokenAmountInput } from '@/hooks/tokens/useTokenAmountInput';
-import { type ExtendedChain } from '@lifi/sdk';
+import type { ExtendedChain } from '@lifi/sdk';
 import { useTranslation } from 'react-i18next';
 import { usePortfolioFormatters } from '@/hooks/tokens/usePortfolioFormatters';
 import { INITIAL_MAX_THRESHOLD_USD, MAX_SELECTABLE_TOKENS } from '../constants';
 import { useAccount } from '@lifi/wallet-management';
-import { checkBalanceWithinRange, getChainMinUsdThreshold } from '../utils';
 import { useTokenFormatters } from '@/hooks/tokens/useTokenFormatters';
-
-export interface DustSummaryValue {
-  selectedBalances: PortfolioBalance<WalletToken>[];
-  nativeToken: ExtendedToken;
-  amount: string;
-  amountUSD: number;
-  address: string;
-}
-
-interface DustFieldDeriveResult {
-  threshold: number | undefined;
-  chainId: number | undefined;
-  isValid: boolean;
-}
-
-const sortByAmountDesc = (
-  a: PortfolioBalance<WalletToken>,
-  b: PortfolioBalance<WalletToken>,
-) => b.amountUSD - a.amountUSD;
-
-const selectTopAddresses = (
-  balances: PortfolioBalance<WalletToken>[],
-): string[] =>
-  [...balances]
-    .sort(sortByAmountDesc)
-    .slice(0, MAX_SELECTABLE_TOKENS)
-    .map((b) => b.token.address);
-
-const createDustFieldDerive = (
-  getValue: (key: string) => unknown,
-): DustFieldDeriveResult => {
-  const threshold = getValue('amountThreshold') as
-    | NumericSelectValue
-    | undefined;
-  const chain = getValue('chain') as ChainSingleSelectValue | undefined;
-
-  return {
-    threshold: threshold?.value,
-    chainId: chain?.selectedChain,
-    isValid: !!(threshold?.value != null && chain?.selectedChain != null),
-  };
-};
+import {
+  checkChainHasBalancesBelowThreshold,
+  computeDustAmounts,
+  createDustFieldDerive,
+  dustSummaryContentEqual,
+  getBalancesFieldValue,
+  getDefaultChainAndBalances,
+  getFilteredBalances,
+  resolveAccountAddress,
+  resolveNativeTokenForChain,
+  selectTopAddresses,
+  sortByAmountDesc,
+  sortChainsByFilteredDustUsdDesc,
+  type DustAmountConverters,
+} from '../dustDomain';
+import type { DustSummaryValue } from '../dustTypes';
 
 interface UseDustFormFieldsParams {
   chains: ExtendedChain[];
@@ -86,53 +53,14 @@ export const useDustFormFields = ({
     usePortfolioFormatters();
   const { toDisplayAmountUSD } = useTokenFormatters();
 
-  const checkChainHasBalancesBelowThreshold = useCallback(
-    (chainId: number, maxUsd: number): boolean =>
-      nonNativeBalances
-        .filter((b) => b.token.chainId === chainId)
-        .some((b) =>
-          checkBalanceWithinRange(b, maxUsd, getChainMinUsdThreshold(chainId)),
-        ),
-    [nonNativeBalances],
-  );
-
-  const getFilteredBalances = useCallback(
-    (chainId: number, maxUsd: number) =>
-      nonNativeBalances
-        .filter((b) => b.token.chainId === chainId)
-        .filter((b) =>
-          checkBalanceWithinRange(b, maxUsd, getChainMinUsdThreshold(chainId)),
-        ),
-    [nonNativeBalances],
-  );
-
-  const computeAmounts = useCallback(
-    (
-      filteredBalances: PortfolioBalance<WalletToken>[],
-      selectedAddresses: string[],
-      token: ExtendedToken,
-    ) => {
-      const toTokenStr = (usd: number) =>
-        toAmountFromPrice(
-          toInputAmount(usd.toLocaleString('fullwide'), usdDecimals),
-          token.priceUSD,
-        );
-
-      const maxAmountUSD = toAggregatedAmountUSD(filteredBalances);
-      const amountUSD = toAggregatedAmountUSD(
-        filteredBalances.filter((b) =>
-          selectedAddresses.includes(b.token.address),
-        ),
-      );
-
-      return {
-        amount: toRawAmount(toTokenStr(amountUSD), token.decimals).toString(),
-        maxAmount: toRawAmount(
-          toTokenStr(maxAmountUSD),
-          token.decimals,
-        ).toString(),
-      };
-    },
+  const amountConverters: DustAmountConverters = useMemo(
+    () => ({
+      toAmountFromPrice,
+      toInputAmount,
+      toRawAmount,
+      toAggregatedAmountUSD,
+      usdDecimals,
+    }),
     [
       toAmountFromPrice,
       toInputAmount,
@@ -142,62 +70,21 @@ export const useDustFormFields = ({
     ],
   );
 
-  const defaultChainAndBalances = useMemo(() => {
-    const chainBalanceSums = new Map<number, number>();
-    const chainBalances = new Map<number, typeof nonNativeBalances>();
+  const checkChainHasBalances = useCallback(
+    (chainId: number, maxUsd: number): boolean =>
+      checkChainHasBalancesBelowThreshold(nonNativeBalances, chainId, maxUsd),
+    [nonNativeBalances],
+  );
 
-    let maxChainId: number | undefined;
-    let maxSum = -Infinity;
+  const getFiltered = useCallback(
+    (chainId: number, maxUsd: number) =>
+      getFilteredBalances(nonNativeBalances, chainId, maxUsd),
+    [nonNativeBalances],
+  );
 
-    for (const balance of nonNativeBalances) {
-      const { chainId } = balance.token;
-
-      if (
-        !checkBalanceWithinRange(
-          balance,
-          INITIAL_MAX_THRESHOLD_USD,
-          getChainMinUsdThreshold(chainId),
-        )
-      ) {
-        continue;
-      }
-
-      const newSum = (chainBalanceSums.get(chainId) ?? 0) + balance.amountUSD;
-      chainBalanceSums.set(chainId, newSum);
-
-      const list = chainBalances.get(chainId) ?? [];
-      list.push(balance);
-      chainBalances.set(chainId, list);
-
-      if (newSum > maxSum) {
-        maxSum = newSum;
-        maxChainId = chainId;
-      }
-    }
-
-    if (maxChainId == null) {
-      return { chainId: undefined, addresses: [] };
-    }
-
-    return {
-      chainId: maxChainId,
-      addresses: selectTopAddresses(chainBalances.get(maxChainId) ?? []),
-    };
-  }, [nonNativeBalances]);
-
-  const computeAccountAddress = useCallback(
-    (chainId: number) => {
-      const selectedChain = chains.find((c) => c.id === chainId);
-      if (!selectedChain) {
-        return undefined;
-      }
-      const account = accounts.find(
-        (a) => a.chainType === selectedChain.chainType,
-      );
-
-      return account?.address;
-    },
-    [chains, accounts],
+  const defaultChainAndBalances = useMemo(
+    () => getDefaultChainAndBalances(nonNativeBalances),
+    [nonNativeBalances],
   );
 
   const computeDustSummary = useCallback(
@@ -208,30 +95,31 @@ export const useDustFormFields = ({
         return undefined;
       }
 
-      const address = computeAccountAddress(chainId);
+      const address = resolveAccountAddress(chainId, chains, accounts);
 
       if (!address) {
         return undefined;
       }
 
-      const balances = getValue('balances') as
-        | BalancesMultiSelectValue
-        | undefined;
+      const balances = getBalancesFieldValue(getValue);
       const selectedAddresses = balances?.selectedAddresses ?? [];
 
-      const token =
-        nativeExtendedTokens.find((t) => t.chainId === chainId) ??
-        fallbackNativeToken;
+      const token = resolveNativeTokenForChain(
+        chainId,
+        nativeExtendedTokens,
+        fallbackNativeToken,
+      );
 
-      const filteredBalances = getFilteredBalances(chainId, threshold);
+      const filteredBalances = getFiltered(chainId, threshold);
       const selectedBalances = filteredBalances.filter((b) =>
         selectedAddresses.includes(b.token.address),
       );
       const amountUSD = toAggregatedAmountUSD(selectedBalances);
-      const { amount } = computeAmounts(
+      const { amount } = computeDustAmounts(
         filteredBalances,
         selectedAddresses,
         token,
+        amountConverters,
       );
 
       const next: DustSummaryValue = {
@@ -243,16 +131,7 @@ export const useDustFormFields = ({
       };
 
       const prev = getValue('dustSummary') as DustSummaryValue | undefined;
-      if (
-        prev &&
-        prev.amount === next.amount &&
-        prev.amountUSD === next.amountUSD &&
-        prev.nativeToken.chainId === next.nativeToken.chainId &&
-        prev.selectedBalances.length === next.selectedBalances.length &&
-        prev.selectedBalances.every(
-          (b, i) => b.token.address === next.selectedBalances[i].token.address,
-        )
-      ) {
+      if (prev && dustSummaryContentEqual(prev, next)) {
         return prev;
       }
       return next;
@@ -260,11 +139,26 @@ export const useDustFormFields = ({
     [
       nativeExtendedTokens,
       fallbackNativeToken,
-      getFilteredBalances,
-      computeAccountAddress,
-      computeAmounts,
+      getFiltered,
+      chains,
+      accounts,
+      amountConverters,
       toAggregatedAmountUSD,
     ],
+  );
+
+  const sanitizeBalancesAfterChainOrThreshold = useCallback(
+    ({ getValue }: { getValue: (key: string) => unknown }) => {
+      const { threshold, chainId, isValid } = createDustFieldDerive(getValue);
+      if (!isValid || threshold == null || chainId == null) {
+        return undefined;
+      }
+      const addresses = selectTopAddresses(getFiltered(chainId, threshold));
+      return addresses.length > 0
+        ? { selectedAddresses: addresses }
+        : undefined;
+    },
+    [getFiltered],
   );
 
   return useMemo(
@@ -299,14 +193,16 @@ export const useDustFormFields = ({
           if (threshold == null) {
             return {};
           }
-          const filteredChains = chains.filter((c) =>
-            checkChainHasBalancesBelowThreshold(c.id, threshold),
+          const filteredChains = sortChainsByFilteredDustUsdDesc(
+            chains.filter((c) => checkChainHasBalances(c.id, threshold)),
+            nonNativeBalances,
+            threshold,
           );
 
           const chainAmounts: Record<number, string> = {};
           let description: string | undefined;
           for (const chain of filteredChains) {
-            const balances = getFilteredBalances(chain.id, threshold);
+            const balances = getFiltered(chain.id, threshold);
             if (balances.length === 0) {
               continue;
             }
@@ -353,7 +249,7 @@ export const useDustFormFields = ({
               sidePanelProps: { availableBalances: [] },
             };
           }
-          const filteredBalances = getFilteredBalances(chainId, threshold);
+          const filteredBalances = getFiltered(chainId, threshold);
           const balanceAmounts: Record<string, string> = {};
           for (const balance of filteredBalances) {
             balanceAmounts[balance.token.address] = toDisplayAmountUSD(balance);
@@ -369,19 +265,11 @@ export const useDustFormFields = ({
         sanitizeOn: [
           {
             watchKey: 'chain',
-            sanitize: ({ getValue }) => {
-              const { threshold, chainId, isValid } =
-                createDustFieldDerive(getValue);
-              if (!isValid || threshold == null || chainId == null) {
-                return undefined;
-              }
-              const addresses = selectTopAddresses(
-                getFilteredBalances(chainId, threshold),
-              );
-              return addresses.length > 0
-                ? { selectedAddresses: addresses }
-                : undefined;
-            },
+            sanitize: sanitizeBalancesAfterChainOrThreshold,
+          },
+          {
+            watchKey: 'amountThreshold',
+            sanitize: sanitizeBalancesAfterChainOrThreshold,
           },
         ],
       }),
@@ -404,19 +292,20 @@ export const useDustFormFields = ({
             return {};
           }
 
-          const token =
-            nativeExtendedTokens.find((t) => t.chainId === chainId) ??
-            fallbackNativeToken;
+          const token = resolveNativeTokenForChain(
+            chainId,
+            nativeExtendedTokens,
+            fallbackNativeToken,
+          );
 
-          const filteredBalances = getFilteredBalances(chainId, threshold);
-          const balances = getValue('balances') as
-            | BalancesMultiSelectValue
-            | undefined;
+          const filteredBalances = getFiltered(chainId, threshold);
+          const balances = getBalancesFieldValue(getValue);
 
-          const { amount, maxAmount } = computeAmounts(
+          const { amount, maxAmount } = computeDustAmounts(
             filteredBalances,
             balances?.selectedAddresses ?? [],
             token,
+            amountConverters,
           );
 
           return { fieldProps: { token, amount, maxAmount } };
@@ -435,12 +324,13 @@ export const useDustFormFields = ({
       nativeExtendedTokens,
       fallbackNativeToken,
       defaultChainAndBalances,
-      checkChainHasBalancesBelowThreshold,
-      getFilteredBalances,
+      checkChainHasBalances,
+      getFiltered,
       toDisplayAggregatedAmountUSD,
       toDisplayAmountUSD,
-      computeAmounts,
+      amountConverters,
       computeDustSummary,
+      sanitizeBalancesAfterChainOrThreshold,
       t,
     ],
   );
